feat: parse in-process with axilog and read FightData everywhere
The cutover. `plugin.rs` now calls the new `parse::parse_log`, which
runs axilog inside the game process and returns a `FightData` -- the
`ReportV1` is dropped inside that function and never retained, which is
the whole memory rationale for this migration. `FightRecord` holds a
`FightData`, `Derived::compute` reads one, and every leaf module and UI
surface follows.

This has to be one commit: nothing between "the leaf modules still speak
Elite Insights" and "everything speaks FightData" compiles.

Three things that are more than a field rename:

Position tracks are ragged. Each track starts at that player's own
first-aware time rounded up to the shared polling grid, and this
fixture's 93 tracks start anywhere from 0ms to 100800ms. The old code
indexed two players' tracks in lockstep on the assumption that EI's
single `pollingRate` implied one grid starting at t=0 for everyone, so
`timeline_distance` was subtracting positions up to a minute and a half
apart, and `lerp_position` drew late joiners wherever someone else's
history happened to sit. `FightData` now carries `poll_ms` and each
track's `track_start_ms` -- a lossless stand-in for the per-sample
timestamps, since axilog's downsampler emits one sample per grid tick
with no gaps -- and every lookup converts a TIME to that track's own
index. Nothing compares two tracks by index any more.

Positions are raw world inches now, not pre-projected pixels, so
`ui/map.rs` projects each `(x, y)` independently through
`FightData::arena`. Verified that axilog's arena world rect and this
crate's own `map_pixel_size` table describe the same region, so
landmarks and tiles still line up.

`healing_available` is read by the Pulse Support subview, which shows
its existing not-available layout rather than the structural zeros a log
without the arcdps healing addon carries. `dist_to_com` stays `Option`
all the way to the card, which renders an em dash rather than a 0.

`self_identify.rs` is deleted: `FightData::self_idx` resolves the
recorder at projection time.

The Elite Insights code is orphaned, not removed -- Task 8 deletes it.
`slim.rs` and its test keep their own copies of the accessors they used
to borrow from `pulse_metrics` so both still compile until then.

+++ src/map/boon_panel.rs
@@ -5,7 +5,7 @@
 //! damage/quickness/alacrity, then utility. Limited list keeps the
 //! row visually scannable.
 
-pub const PANEL_BOON_ORDER: &[i64] = &[
+pub const PANEL_BOON_ORDER: &[u32] = &[
     740,   // Might
     725,   // Fury
     717,   // Protection
@@ -21,7 +21,7 @@ pub const PANEL_BOON_ORDER: &[i64] = &[
 ];
 
 /// Short display name (tooltip text).
-pub fn boon_name(id: i64) -> &'static str {
+pub fn boon_name(id: u32) -> &'static str {
     match id {
         740 => "Might",
         725 => "Fury",
